obtimized code

+++ server/tcp.go
@@ -12,7 +12,7 @@ import (
 type TcpService struct {
 	connMax   int
 	connCount int
-	curFD     int
+	curFD     uint64
 	listener  net.Listener
 	locker    sync.Mutex
 	isClosed  bool
